pyproject.toml
[project]
name = "easymdm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ankit K", email = "ankit48365@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "pandas>=2.3.1",
    "pyaml>=25.7.0",
    "python-levenshtein>=0.27.1",
    "recordlinkage>=0.16",
]

[project.scripts]
easymdm = "easymdm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
